--- src/main/java/app/entities/User.java
@@ -1,45 +1,0 @@
-package app.entities;
-
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.mindrot.jbcrypt.BCrypt;
-
-import java.util.Set;
-
-@Data
-@NoArgsConstructor
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @Basic(optional = false)
-    @Column(unique = true, nullable = false)
-    private String username;
-    private String password;
-    @ManyToMany(fetch = FetchType.EAGER)
-    private Set<Role> roles;
-
-    @Builder
-    public User(String username, String password, Set<Role> roles) {
-        this.username = username;
-        this.password = BCrypt.hashpw(password, BCrypt.gensalt());
-        this.roles = roles;
-    }
-
-    public boolean verifyPassword(String password) {
-        return BCrypt.checkpw(password, this.password);
-    }
-
-    public void addRole(Role role) {
-        roles.add(role);
-        role.getUsers().add(this);
-    }
-
-    public void removeRole(Role role) {
-        roles.remove(role);
-        role.getUsers().remove(this);
-    }
-}
