--- src/main/java/app/exceptions/PasswordValidationException.java
@@ -1,8 +1,0 @@
-package app.exceptions;
-
-public class PasswordValidationException extends RuntimeException {
-
-    public PasswordValidationException(String message) {
-        super(message);
-    }
-}
