--- src/main/java/app/routes/Routes.java
@@ -1,11 +1,0 @@
-package app.routes;
-
-import io.javalin.apibuilder.EndpointGroup;
-
-public class Routes {
-
-    public EndpointGroup getAPIRoutes() {
-        return () -> {
-        };
-    }
-}
